Remove redundant switch statements
Without more context, it appears that the switch statements don't do anything. Perhaps someone changed this from configurable to hardcoded at some point and didn't clean up afterwards. Another possibility is that this is a bug and there is a config somewhere that should be used as the value in the switch statement.